# src/swagger/swagger.json
{"openapi":"3.1.0","servers":[{"url":"http://localhost:80","description":"게시판"}],"info":{"title":"veightcorp API","description":"브이에이트코프 과제 API 문서입니다.","version":"1.0.0","license":{"name":"UNLICENSED"}},"paths":{"/aws":{"post":{"tags":["AWS"],"parameters":[],"responses":{"201":{"description":"S3 이미지 업로드 여부를 반환합니다.","content":{"application/json":{"schema":{}}}}},"summary":"토큰이 만료되었다면 에러를 쓰로잉 합니다","description":"토큰이 만료되었다면 에러를 쓰로잉 합니다."}},"/post":{"get":{"tags":["Post"],"parameters":[{"name":"query","in":"query","schema":{"$ref":"#/components/schemas/PostParam.Get.Search"},"description":"","required":true}],"responses":{"200":{"description":"게시글을 조회합니다.","content":{"application/json":{"schema":{}}}}},"summary":"주어진 쿼리 파라미터에 따라 다른 결과를 반환합니다","description":"주어진 쿼리 파라미터에 따라 다른 결과를 반환합니다."},"post":{"tags":["Post"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostBody.Post.Create"}}},"required":true},"responses":{"201":{"description":"게시글 작성 여부를 반환합니다.","content":{"application/json":{"schema":{}}}}},"summary":"토큰이 만료되었다면 에러를 쓰로잉 합니다","description":"토큰이 만료되었다면 에러를 쓰로잉 합니다."},"patch":{"tags":["Post"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostBody.Patch.Update"}}},"required":true},"responses":{"200":{"description":"게시글 수정 여부를 반환합니다.","content":{"application/json":{"schema":{}}}}},"summary":"토큰이 만료되었다면 에러를 쓰로잉 합니다","description":"토큰이 만료되었다면 에러를 쓰로잉 합니다."},"delete":{"tags":["Post"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostBody.Delete.Remove"}}},"required":true},"responses":{"200":{"description":"게시글 삭제 여부를 반환합니다.","content":{"application/json":{"schema":{}}}}},"summary":"토큰이 만료되었다면 에러를 쓰로잉 합니다","description":"토큰이 만료되었다면 에러를 쓰로잉 합니다."}},"/post/{id}":{"get":{"tags":["Post"],"parameters":[{"name":"id","in":"path","schema":{"type":"string"},"description":"","required":true}],"responses":{"200":{"description":"게시글의 상세정보를 반환합니다.","content":{"application/json":{"schema":{}}}}}}},"/reply":{"post":{"tags":["Reply"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReplyBody.Post.Create"}}},"required":true},"responses":{"201":{"description":"댓글, 대댓글 작성 여부를 반환합니다.","content":{"application/json":{"schema":{}}}}},"summary":"토큰이 만료되었다면 에러를 쓰로잉 합니다","description":"토큰이 만료되었다면 에러를 쓰로잉 합니다."},"patch":{"tags":["Reply"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReplyBody.Patch.Update"}}},"required":true},"responses":{"200":{"description":"댓글, 대댓글 수정 여부를 반환합니다.","content":{"application/json":{"schema":{}}}}},"summary":"토큰이 만료되었다면 에러를 쓰로잉 합니다","description":"토큰이 만료되었다면 에러를 쓰로잉 합니다."},"delete":{"tags":["Reply"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReplyBody.Delete.Remove"}}},"required":true},"responses":{"200":{"description":"댓글, 대댓글 삭제 여부를 반환합니다.","content":{"application/json":{"schema":{}}}}},"summary":"토큰이 만료되었다면 에러를 쓰로잉 합니다","description":"토큰이 만료되었다면 에러를 쓰로잉 합니다."}},"/user/login":{"post":{"tags":["User"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBody.Post.Login"}}},"required":true},"responses":{"201":{"description":"로그인에 성공하면 쿠키에 토큰정보를 저장하고, 유저 정보를 반환합니다."}}}},"/user/login/token":{"post":{"tags":["User"],"parameters":[],"responses":{"201":{"description":"토큰이 유효하다면 유저 정보를 반환합니다."}},"summary":"액세스 토큰이 만료되었을 경우, 새로운 액세스 토큰을 반환하고, 쿠키를 갱신합니다","description":"액세스 토큰이 만료되었을 경우, 새로운 액세스 토큰을 반환하고, 쿠키를 갱신합니다.\n\n리프레시 토큰이 만료되었을 경우, 쿠키에 담긴 토큰정보를 삭제하고, 재 로그인을 유도합니다."}},"/user/sign":{"post":{"tags":["User"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBody.Post.SignUp"}}},"required":true},"responses":{"201":{"description":"회원가입에 성공 여부를 반환합니다.","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"PostParam.Get.Search":{"type":"object","properties":{"title":{"type":"string","maxLength":100},"writer_email":{"type":"string","format":"email"},"withBy":{"oneOf":[{"const":"year"},{"const":"month"},{"const":"week"}]},"createdAt":{"oneOf":[{"const":"asc"},{"const":"desc"}]},"viewAt":{"oneOf":[{"const":"asc"},{"const":"desc"}]}}},"PostBody.Post.Create":{"type":"object","properties":{"title":{"type":"string","maxLength":100},"category":{"oneOf":[{"const":"q-and-a","maxLength":10},{"const":"notice","maxLength":10},{"const":"man-to-man","maxLength":10}]},"contents":{"type":"string"}},"required":["title","category","contents"]},"PostBody.Patch.Update":{"type":"object","properties":{"post_id":{"type":"string","maxLength":32},"category":{"oneOf":[{"const":"q-and-a","maxLength":10},{"const":"notice","maxLength":10},{"const":"man-to-man","maxLength":10}]},"title":{"type":"string","maxLength":100},"contents":{"type":"string"}},"required":["post_id","category"]},"PostBody.Delete.Remove":{"type":"object","properties":{"post_id":{"type":"string","maxLength":32},"category":{"oneOf":[{"const":"q-and-a","maxLength":10},{"const":"notice","maxLength":10},{"const":"man-to-man","maxLength":10}]}},"required":["post_id","category"]},"ReplyBody.Post.Create":{"type":"object","properties":{"contents":{"type":"string","maxLength":200},"detail_id":{"type":"string","maxLength":32},"comment_id":{"type":"string","maxLength":32}},"required":["contents","detail_id"]},"ReplyBody.Patch.Update":{"type":"object","properties":{"reply_id":{"type":"string","maxLength":32},"contents":{"type":"string","maxLength":200}},"required":["reply_id","contents"]},"ReplyBody.Delete.Remove":{"type":"object","properties":{"reply_id":{"type":"string","maxLength":32}},"required":["reply_id"]},"UserBody.Post.Login":{"type":"object","properties":{"email":{"type":"string","format":"email"},"pass":{"type":"string","maxLength":100}},"required":["email","pass"]},"UserBody.Post.SignUp":{"type":"object","properties":{"email":{"type":"string","format":"email"},"pass":{"type":"string","maxLength":100}},"required":["email","pass"]}}},"tags":[{"name":"AWS","description":"Private"},{"name":"Post","description":"Public"},{"name":"Reply","description":"Private"},{"name":"User","description":"Public"}],"x-samchon-emended":true}